Format webrtc_internals.html
Run a formatter on webrtc_internals.html, and nudge it in a few
places, so as to produce a maintainable result.

This CL does not affect any material changes[*].
Those is left for subsequent CLs, where the diff
produced by editing the code would be more obvious,
following the formatting that is done here.

[*] However, some errors in the HTML, found by the formatter,
are fixed. These include:
- Missing quotations marks.
- Superfluous quotation marks.

Bug: 417262733
Change-Id: I4b3382f1ea3c7f5c65e8edfab38af74968f1271b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6533311
Commit-Queue: Elad Alon <eladalon@chromium.org>
Reviewed-by: Philip Eliasson <philipel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459271}