Avoid visible anchor in WebRTC Internals
WebRTC Internals makes use of an anchor (<a> tag) behind the scenes
to facilitate a download. Unfortunately, this anchor can sometimes
become visible as a linkable empty character ("space").

This CL fixes the issue by replacing the in-DOM anchor with
one that is produced programmatically and is never inserted
into the DOM.

Screenshot of original issue:
https://screenshot.googleplex.com/8L76RwnZNuBkysC

Bug: 417262733
Change-Id: Id8c953ef624182c5ecf04579012fc42c87930a7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6533314
Reviewed-by: Philip Eliasson <philipel@chromium.org>
Commit-Queue: Elad Alon <eladalon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459274}