Log duplicate network requests in HttpNetworkTransaction
Add code to net::HttpNetworkTransaction to track all GET requests sent
over the network in the last 10 seconds. For those that are duplicates,
the time since the previous identical request is logged in a UMA
histogram.

Requests are considered identical purely based on URL match as this is
what the server sees.

We only take an interest in GET requests, as POST requests having the
same URL is not unusual, and other HTTP methods are not interesting.

The feature is protected by a base::Feature, "LogDuplicateRequests", so
that it can be enabled only on Dev and Canary channels. It should not be
enabled for stable users as it has some performance overhead.

This change is not intended to be permanent. It can be reverted once the
root cause of https://crbug.com/413557424 is found.

As a mostly-unrelated change, replace
`IsGoogleHostWithAlpnH3(url_.host())` with
`IsGoogleHostWithAlpnH3(url_.host_piece())` throughout
http_network_transaction.cc, to avoid creating a `std::string` every
time. This part should not be reverted.

BUG=413557424

Change-Id: I6ac8fcfc5bed7bf7207fffd8f815888a40739e0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6532455
Reviewed-by: Nidhi Jaju <nidhijaju@chromium.org>
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459280}