[UNO-d] Update sync promo description in profile menu
It updates the profile menu history sync opt-in promo profile menu text
and button label when `EnableHistorySyncOptinExpansionPill` is enabled.

This is the revision of crrev.com/c/6513036. Agreed with the PM that the
updated history sync opt-in text should be shown whenever the profile
menu is opened (it doesn't have to be triggered from the pill).

Bug: 415748973
Change-Id: I02132087f006db3fdd2c2044ad420f1ae24107f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6533983
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Ernest Nguyen Hung <ernn@google.com>
Cr-Commit-Position: refs/heads/main@{#1459340}