printing: fix crash when custom margins are set via print preview dialog
Previously, support for custom margins was added for chrome.printing
API. Users of the API are allowed to use only those margins, which are
announced by the printer. As a result, all those custom margins are
backwards convertible from microns to PWG units.

However, setting custom margins via the print preview dialog is not
restricted to only supported margins. Instead, preprocessing of the
document happens, but when the print job is setup, default margins must
be used (that was the original behaviour). After the above mentioned
change, those custom margins started to be used when print jobs are
setup. The helper that converts margins expects that those are backwards
convertible. But given unsupported custom margins were used, it resulted
in the CHECK fired.

This patch restores the initial behaviour for the print preview dialog.
If margins are custom (it's unknown if they come from the print preview
dialog or the chrome.printing API actually), try to convert them to PWG
units first and if it fails, use default margins obtained from the
printable area.

Bug: 308709702, 316999874, 416514812
Change-Id: Id592d7d57e35bf63960b6b31c3e1aa94878a5651
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6530081
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1459345}