[CDP] Avoid race in Page.captureScreenshot
Problem: in rare circumstances Page.captureScreenshot stalls.
This happens because WidgetBase::ForceRedraw never returns
due to compositor frame never being presented. Screenshot forces
redraw to ensure that fresh content from the page is visible.

This change follows the advice in
https://issues.chromium.org/u/1/issues/377715191#comment12
to allocate a new LocalSurfaceId and rely on the visual
properties synchronization to repaint onto the new surface.
This way, when requesting a copy from the new surface we
are guaranteed to receive the fresh content from the renderer,
especially if we forced redraw in the renderer beforehand.

Since this new behavior could break in the wild due to
unforeseen reasons, it is guarded by the feature for now.
We'll try running in the wild with the feature enabled and
gather feedback. Long term plan is to enable the feature
and remove old code path.

Bug: 377715191

Change-Id: I23ddf54fee0efc1dc6d05dec6f6fc24be23ae1ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6519891
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Dmitry Gozman <dgozman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459358}