Fix dangling observer pointer in DeviceDisablingManager
The observer (DeviceDisabledScreen) never unregistered itself,
causing a dangling pointer after the screen has been dismissed.
This could previously not happen but since the introduction of
the DeviceRestrictionSchedule in M132 this screen will be dismissed
when the restriction schedule is over.

The actual crash happened when the `OnRestrictionScheduleMessageChanged`
event arrived after the device was released, which doesn't look like
a very uncommon event since the timer that sends this event is seemingly
never cancelled (see
`DeviceRestrictionScheduleControllerImpl::StartMessageUpdateTimer`).

Fixed: 414743653
Change-Id: Iadd9b6483111b6bae9ecd79b05ce0964f05ae0e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6498207
Reviewed-by: Ahmed Mehfooz <amehfooz@chromium.org>
Reviewed-by: Danila Kuzmin <dkuzmin@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Jeroen Dhollander <jeroendh@google.com>
Cr-Commit-Position: refs/heads/main@{#1459362}