Supervised User Metrics Service is observer of Supervised User Service
The metrics service is merely an observer to changes in the main
service. This change simplifies the architecture of the supervised user
stack by condensing the metrics logic in a dedicated service.

Minor: removes test-only interfaces in favour of direct values, which
are recommended (tott/746, tott/741).

Change-Id: I4368af428df661efad4a10eef8d13a4d93887d52
Bug: 416249869, 413654322
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6534165
Reviewed-by: Liza Bipin <mlbipin@google.com>
Commit-Queue: Tomek Jurkiewicz <tju@google.com>
Cr-Commit-Position: refs/heads/main@{#1459376}