[Shared Tab Groups] Add version field to the specifics
This CL adds a version field to the SharedTabGroupDataSpecifics.
This is intended to be used in conjunction with newly added
fields to the specifics in future as future code will be able to
determine whether the field was written by a client that supported
it or just carried over by an old client which didn't understand the
field.  In this CL, all the write paths are covered so that the
proto will always contain the version field.
The read path isn't implemented and left for future work when a
case arises.

Bug: 372192105
Change-Id: I6f987ddbf18dc3c11b03a9544bc2b2c4addec590
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6533621
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Cr-Commit-Position: refs/heads/main@{#1459378}