Add last_successful_origin_ to FrameTreeNode.
When monitoring the number of same-site fenced frames in the viewport,
as outlined in https://crrev.com/c/6420184, there's a question of what
site an error frame should be monitored under. Error frames have an
opaque origin, meaning a fenced frame could forcibly navigate itself to
an error page to escape monitoring for its current site.

This CL introduces the `last_successful_origin_` field on FrameTreeNode,
which is only updated when a navigation commits a non-error page. This
field can be used in cases like the above, where the error page needs to
be connected to the site which was previously rendered in the frame.

However, the *currently committed origin* in the frame should still be
used for security-sensitive decisions, and the getter for the new field
is documented to state as such.

Change-Id: I865f67e53bd26f99d256d5140fbda5f90e4bb9ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6499627
Commit-Queue: Andrew Verge <averge@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459386}