[GAR] Improve editor "required" fields accessibility experience
This commit improves the accessibility of "required" fields in Autofill
editors for Android. Previously, a simple '*' was appended to the field
label to indicate it was required, which is not announced by screen
readers.

Now, for accessibility purposes, the full string "required" is appended
to the field's content description, making it explicitly announced to
screen reader users. The visual hint ('*') remains unchanged.

Additionally, the "Required field" message in the footer is now marked
as android:importantForAccessibility="no" to prevent redundant
announcements, as the "required" status is now conveyed directly on the
field itself.

** UI video **
http://screencast/cast/NTcwMTYwNzI3NzcyMzY0OHxiNTI0YmUyNC05YQ

Bug: 401373578
Change-Id: I9d7398b9996c6a9adef652b81b32e723c002e70b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6533102
Commit-Queue: Piotr Kotynia <piotrkotynia@google.com>
Reviewed-by: Friedrich Hauser <friedrichh@chromium.org>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1459390}