[ios] Store reference to owning Browser in BrowserUserData<T>
Instead of having each BrowserUserData<T> subclasses managed the
pointer to the owning Browser, directly store it in the base class.
This is possible because now all sub-class pass the pointer to the
Browser to BrowserUserData<T> constructor.

The subclasses will be changed in followup CLs to use the pointer
that is stored in the base class and stop managing it themselves.

Bug: 409299519
Change-Id: Ib0c5ff68d61cb4ebc308523c8a0c186cad223477
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6519772
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459393}