[lens-rearch][19/n] Move side panel closing logic to search controller
This CL moves the OnSidePanelWillHide and OnSidePanelHidden logic from
the overlay controller to the search controller. This is one more step
needed to decouple the overlay from the side panel. In doing this, I was
also able to simplify the overlay close calls, to give more control of
the async closing to the search controller. This also allowed me to
delete CloseUIAsync and CloseUISync from the overlay controller and
instead rely on TriggerOverlayCloseAnimation and a single CloseUI
method.

Bug: 404941800
Change-Id: Ib7e681151cb3a7317d86ec20ea1c6b3ddba45e28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6535654
Reviewed-by: Juan Mojica <juanmojica@google.com>
Commit-Queue: Duncan Mercer <mercerd@google.com>
Cr-Commit-Position: refs/heads/main@{#1459407}