Make non-structured attributes field-type-independent in reading/writing
This change is motivated by cases where:
1) A field has both Autofill and Autofill AI predictions.
2) The Autofill prediction is wrong.
3) The Autofill AI prediction is not one of a structured type.

An example of such a case is in the linked bug.

Note that this change:

A) Gives priority to Autofill AI server predictions over Autofill
   Heuristics, given the priority they already have in suggestion
   generation.
B) Gives priority to Autofill AI server predictions over the HTML
   autocomplete attribute, same as above.
C) Would disable regular autofill for (union of):
   C1) Users that have an entity saved.
   C2) Fields were both predictions exist and Autofill AI is wrong.
D) Allows filling a field with info that doesn't match
   `AutofillField::overall_type_`.

We are relying on the gradual increase of users satisfying C1 and on
the belief that C2 should not be occurring too often.

Fixed: 415960797
Change-Id: Icd39a5acb16f880a51549a8fd8257461be679b31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6540371
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1459441}