[TreesInViz] Clean up cc.mojom.ElementId type traits
cc::ElementId has two reserved internal values, kInvalidElementId
and kDeletedElementId, that are not expected to be constructed explicitly
by clients.

We need to be able to serialize these values over Mojo. The existing code
handles the kInvalidElementId case via automatic nullability, but this
is confusing, and also misses handling of the kDeletedElementId case.

Instead, remove the nullability and handle both special values
explicitly.

Bug: 40902503
Change-Id: I54d2eb31a062bee1787cb8145eee035a15763d4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6535304
Commit-Queue: Victor Miura <vmiura@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Fred Shih <ffred@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459449}