[Minimal UI] Add padding explicitly for status bars
The web app header layout needs to know if the caption bar insets also
include any status bar insets so that we know where the best place to
position web app header controls is.

Prior to this CL, all we had was a single Rect representing the caption
bar insets but we didn't know how much of that was the status bar. An
assumption was made that the app header would always be 40dp and any
caption bar insets larger than 40dp would represent the status bar.

This CL exposes an AppHeaderObserver API to observe header controls top
offset changes to retrieve the height of a possibly overlapping status
bar. This can be used by the web app header layout to calculate the best
position for header controls.

Bug: 409497564
Change-Id: I68987ec6ba12009f1193172b66543b0d5f0ac4d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6496039
Reviewed-by: Aishwarya Rajesh <aishwaryarj@google.com>
Reviewed-by: Vlad Korotkevich <vkorotkevich@google.com>
Commit-Queue: Adam Raine <asraine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459482}