[B4B] Fix getting default parent for managed nodes
Some operations are allowed for managed nodes, such as opening the
bookmark manager after selecting a managed node through the context
menu.

The code previously did not allow that since it assumed this was only
used to get the parent to create new nodes, which is not allowed for
managed bookmarks.

Add a method specific for managed nodes in the MergedService to
expose that this is an accepted operation but that it should be
treated differently than the regular nodes.

Fixed: 414129879
Change-Id: I4771a41cbc53cba1ce82c6eb46e44e7da3c4eaf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6540051
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1459515}