Changes to WindowTreeSynchronizer/LayerTreeSynchronizer
1. To ensure the layer tree synchronization always uses the current root
layer, do not persist the root layer upon initialization. Retrieve it
every time instead, as the underlying root layer `object` might be
changed. (See b:413577648)

2. For overview item, whenever a mouse is pressed, we start drag event.
However, in some instances, the window drag is never ended. For example,
when the user clicks the overview item to select a window and end the
overview.

In order to have a correct layer tree state, layer trees must be
restored to state before the drag stated. Otherwise it can cause bugs.
(See b:413577648)

3. In crrev.com/c/5486579, the original cached state of the layer tree
was reset in order to not restore layer tree to a stale state. The fix
tried to deal with side effect of dragging an overview item and not
restoring the layer tree to pre-drag state. A fix was introduced in
crrev.com/c/6473585, after which reseting the cached state is not
necessary.

Bug: b:416789348, b:413577648
Change-Id: I16c3ca5a539e9ab6d992a712ab9e4940f13492a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6534063
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Zoraiz Naeem <zoraiznaeem@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459522}