[declutter] Support delayed archived tabmodel when closing tabs
When closing all tabs, it's assumed that the archived tabmodel is
loaded. This isn't the case, especially after crrev.com/cl/6499064
which delays the archived tabmodel even more.

This CL accounts for the possibility that the archived tabmodel isn't
loaded when a user closes all their tabs. It no-ops this case, and
leaves the archived tabs intact.

Bug: 417254184
Change-Id: Icac1d4783ad34d9f98a641145ebffb0a2f228662
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6535109
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Brandon Wylie <wylieb@google.com>
Cr-Commit-Position: refs/heads/main@{#1459537}