Correctly configure transparent TrayBubbleView
Background: A transparent TrayBubbleView is used as a container for
hosting content.

Reasons for the change: To ensure a transparent tray bubble after the
changes in crrev.com/c/6501740, the bubble background color must be
explicitly set to transparent. This is because that change will cause
the client view to always be painted to a layer. Currently, this is not
a problem as TrayBubbleView does not paint its client view to a layer,
so a background on the client view is not painted. Setting the
background to transparent is a prerequisite for crrev.com/c/6501740 to
avoid painting a default background.

Also a transparent bubble should not apply rounded corners, since the
contents of a transparent TrayBubbleView handles its own rounded
corners.

Refactor: Instead of using TrayBubbleView::InitParams to specify the
radius BubbleBorder, we now use the BubbleDialogDelegate to configure
the bubble rounded corners.

[1]
https://source.chromium.org/chromium/chromium/src/+/main:ui/views/bubble/bubble_dialog_delegate_view.h;l=630;drc=4a1f2646209c5ecc0aae120e3654feb6d3f393f3

Bug: b:414655934
Change-Id: I5ee4724dbc5184a12eb7be6889ce7d026e434543
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6541768
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Zoraiz Naeem <zoraiznaeem@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459577}