Clean up and align SelectionController APIs.
The SelectionController was for some time considered as a nested class
or delegate to be used with the RecyclerView in order to solve some
dependency issues. Different names for similar functions were chosen
intentionally to ensure correct methods are called on correct objects.

Some of the dependency issues may not require resolution at this time
(e.g. evaluating if items can be selected based on the property model
values), as the implementation is in "good enough" shape.

Bug: 344930378
Change-Id: Ie74c463813f8b78b7ce0505ececd7a6d41d01f05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6538166
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/main@{#1459579}