[Tab Features] Migrate resource usage to TabFeatures
This changes the resource usage tab helper from a WebContents helper to
a Tab Feature, as per the request in the attached bug.

A number of references and tests had to be modified or moved, but
overall the logic has not changed and things appear to be behaving
themselves.

Bug: 401033983
Change-Id: I72be7539b09bd900e8c5e3157654b0cd57f13f2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6541067
Reviewed-by: Steven Luong <stluong@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459587}