[Gap Decorations]: Parse column-rule shorthands
This CL augments the parsing logic for `column-rule` to support the new
grammar [1] for gap decorations. According to the new grammar,
`column-rule` can now support list of <gap-rule>s and a <gap-rule> can
be a regular or repeated <gap-rule>.

For the parsing logic, the `ConsumeGapDecorationsRuleShorthand` is the
top-level function responsible for parsing and hydrating the width,
style, and color properties associated with gap decorations. It handles
both repeated and non-repeated values.

To get the computed value as specified, each property is iterated over,
and a corresponding <gap-rule> is constructed at each index.

Subsequent CLs will see us implement serialization,`row-rule` and `rule`
properties.

[1]: https://drafts.csswg.org/css-gaps-1/#propdef-column-rule

Bug: 357648037
Change-Id: Ifd73efe7b3f604e2e53cabb68550cf7241c26e44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6423813
Commit-Queue: Sam Davis Omekara <samomekarajr@microsoft.com>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1459604}