[Mac PWAs] Allow chrome://flags to override ad-hoc signed policy
The ad-hoc signed app shim enterprise policy exists for enterprise
admins to prevent breakage to their users. It doesn't exist to
limit what users can do themselves though. Since the goal is to
eventually remove the policy, allowing chrome://flags to override
the policy will make it easier for admins to test the behavior when
ad-hoc signed shims are enabled, even if there is an enterprise
policy for the opposite.

Bug: 417456922
Change-Id: Ic2287f3aa3e9bdfdef46fcf6c1333419688ff858
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6542486
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459605}