De-duplicate speculative decode requests earlier
Prior to this CL, a redundant speculative decode request would
progress all the way through creating and scheduling a decode task
on the compositor thread. Only when the decode task actually ran
would it notice that the image already had decoded data available.

This CL annotates the blink::ImageResource once a speculative decode
has been attempted, so that subsequent requests for the same resource
can avoid the extraneous thread hops and task scheduling. This could
be sub-optimal if we actually need to decode a single image using
different sets of parameters, but that seems like a rare and
unusual situation.

Bug: chromium:378623805
Change-Id: I81b60540742fce8b4c3578db598a0ba0d14fcdb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6537927
Commit-Queue: Stefan Zager <szager@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459654}