[browser] Update RestoreNonEditableTabstrip to use real Browser
Currently ExtensionSessionsTest.RestoreNonEditableTabstrip creates a
semi-initialized Browser by instantiating a TestBrowserWindow (which
skips initialization of the BrowserView and BrowserWidget). This
creates a Browser instance that is not fully representative of a real
browser and improperly initializes the test instance.

As part of the effort to deprecate TestBrowserInstances in such cases
this CL updates this test to use a fully-initialized Browser
instance.

Bug: 413168662
Change-Id: I333d7906ab0de44cc30d49129f57aa84f234dcb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6535002
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Commit-Queue: Tom Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459673}