[sync] Plumb username in Java integration tests with fake identity
Before this patch, tests exercised the account ID hardcoded in Java,
differently to what desktop tests do and ignoring the `username`
parameter passed to `SyncSigninDelegateAndroid::SigninFake()`.

In this patch, Android tests are made more similar to the desktop one by
instead honoring the `username` parameter. When doing so, as per
reviewer feedback in this patch, `signin::GetTestGaiaIdForEmail()` is
adopted also on Android, which means that
`SyncServiceImplHarness::GetGaiaIdForDefaultTestAccount()` can be
simplified.

Bug: 40165479
Change-Id: I6778ab9e3eead9a6d1c061f0b4570e7097685c38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6533978
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1459685}